+++ DK/DKFramework/Private/CocoaTouch/Window.h
@@ -47,7 +47,7 @@ namespace DKFramework::Private::iOS
 
         void SetOrigin(DKPoint) override;
         void Resize(DKSize, const DKPoint* optionalOrigin) override;
-        double ContentScaleFactor() const override;  // logical coords by pixel ratio.
+        float ContentScaleFactor() const override;  // logical coords by pixel ratio.
 
         void SetTitle(const DKString& title) override;
         DKString Title() const override;
