+++ DK/DKFramework/Private/CocoaTouch/Window.mm
@@ -262,7 +262,7 @@ void Window::Resize(DKSize s, const DKPoint* pt)
 	});
 }
 
-double Window::ContentScaleFactor() const
+float Window::ContentScaleFactor() const
 {
 	return DispatchSyncOnMain([&]()->CGFloat {
 		return view.contentScaleFactor;
